+++ src/entities/item.hpp
@@ -12,6 +12,7 @@
 #include "../components/RigidBody.hpp"
 #include "../constants.hpp"
 #include "../ecs/ECSManager.hpp"
+#include "../engine/types/Vec2f.hpp"
 #include "../engine/types/Vec2i.hpp"
 
 #define ITEMS_SPRITESHEET_Y 1000
@@ -29,7 +30,7 @@ struct InteractableItem {
 };
 
 // instantiate item on map (not in some inventory) with tile values, not pixel values.
-Entity instantiateItemEntity(ECSManager &ecs, const Vec2i position, const Vec2i spriteSheetPos)
+Entity instantiateItemEntity(ECSManager &ecs, const Vec2f position, const Vec2i spriteSheetPos)
 {
 	Entity item = ecs.addEntity();
 
@@ -43,7 +44,7 @@ Entity instantiateItemEntity(ECSManager &ecs, const Vec2i position, const Vec2i 
 	return item;
 }
 
-Entity instantiateCollectableItemEntity(ECSManager &ecs, const Vec2i position, CollectableItem collectableItem)
+Entity instantiateCollectableItemEntity(ECSManager &ecs, const Vec2f position, CollectableItem collectableItem)
 {
 	Entity item = instantiateItemEntity(ecs, position, collectableItem.spriteSheetPos);
 	ecs.addComponent(item, std::move(collectableItem.interactable));
@@ -52,7 +53,7 @@ Entity instantiateCollectableItemEntity(ECSManager &ecs, const Vec2i position, C
 	return item;
 }
 
-Entity instantiateInteractableItemEntity(ECSManager &ecs, const Vec2i position, InteractableItem interactableItem)
+Entity instantiateInteractableItemEntity(ECSManager &ecs, const Vec2f position, InteractableItem interactableItem)
 {
 	Entity item = instantiateItemEntity(ecs, position, interactableItem.spriteSheetPos);
 	ecs.addComponent(item, std::move(interactableItem.interactable));
